fix(cid): separates cid from stringUtils - othrwise the TextDecoder issue gets triggered upon testing
