+++ src/utils/stringUtils.ts
@@ -1,25 +1,3 @@
-import CID from 'cids'
-
-export const isEmpty = (
-  text: string | unknown,
-): boolean => !text || !String(text).trim()
-
-export const validateCID = (
-  cid: string,
-  errorHandle?: (e: Error) => unknown,
-): boolean => {
-  try {
-    CID.validateCID(new CID(cid))
-
-    return true
-  } catch (e) {
-    if (errorHandle) {
-      errorHandle(e)
-    }
-    return false
-  }
-}
-
 export const getURLParamByName = (
   url: string, param: string,
 ): string | undefined => {
@@ -27,3 +5,7 @@ export const getURLParamByName = (
   const match = url.match(rgx)
   return match?.find((_, i) => i === 1)
 }
+
+export const isEmpty = (
+  text: string | unknown,
+): boolean => !(text && String(text).trim())
